refactor(m9.4): remove AccountManager static facade + migration guide
Closes the M9 milestone (#270) and audit finding F8 (#213) by deleting
the 17-method static facade that was deprecated in 0.2.7. The instance
API (`new AccountManager(options?)` + `harness.runtime.accountManager`
+ `harness.accounts.<label>`) is now the only surface; calls to the
former static methods fail with `TypeError`.

Per MAINTENANCE.md §6 this ships in 0.3.0 once the one-minor-release
deprecation window elapses (earliest 2026-06-22). The commit lives on
the branch locally until then; push + release PR happen post-window.

Changes
-------

- `core/AccountManager.ts`: deletes the 17 static methods, the
  `__defaultManager` singleton, the `__warnedMethods` Set, the
  `__warnDeprecated` helper, and the `_resetDeprecationWarnings`
  test-only escape hatch. Class JSDoc rewritten — no more
  "Two API surfaces are exposed" / "Static facade (deprecated)"
  paragraphs. Net file shrinks 627 → 445 lines.

- `core/__tests__/AccountManager.deprecation.test.ts`: deleted —
  pinned warning behavior that no longer exists.

- `core/__tests__/AccountManager.global-state.test.ts`: renamed to
  `AccountManager.instance-isolation.test.ts`. The F8 pinning block
  (which captured the now-removed static-state contract) is gone;
  the 4 per-instance isolation tests are preserved verbatim.

- `core/__tests__/AccountManager.test.ts`: every static call replaced
  with a per-describe-block `mgr = new AccountManager()` in
  `beforeEach`. The save/clear/load round-trip tests now use two
  instances (`mgr` + `restored`) to prove the round-trip works across
  boundaries, not just on the same in-memory pool.

- `__tests__/runtime.test.ts`: dropped the now-redundant
  `AccountManager.clearPool()` calls from beforeEach/afterEach —
  the runtime owns its own AccountManager instance per invocation.

- `docs/upgrading/0.3.0.mdx` + `upgrading/meta.json`: NEW migration
  guide. Covers the common labeled-account case, advanced
  direct-manager calls, standalone (non-Harness) usage, per-instance
  pool semantics, the `process.chdir` behavior change, and a
  4-question FAQ. Wired into the docs sidebar via the top-level
  meta.json (new "Upgrading" category between "API Reference" and
  "Community").

- `CHANGELOG.md`: renames the previous `[Unreleased]` block to
  `[0.2.7] - 2026-05-22` (catching up develop with main's release
  rename, missed in the release/0.2.7 cycle). Opens a fresh
  `[Unreleased]` with `### Breaking`, `### Removed`, `### Changed`,
  `### Docs` subsections documenting the M9.4 work.

- `ROADMAP.md`: backfills the M9 milestone section after M8 with the
  standard closeout table (M9.1 → M9.4 sub-PRs). M9 closes audit
  finding F8 across the 0.2.7/0.3.0 release boundary.

Verification
------------

- `pnpm build:movehat`: clean
- `pnpm test`: 49 files / 551 passing (was 553; -2 from the
  deprecation test deletion + F8 block deletion, +0 net since the
  new instance-isolation tests are the same 4 as before, just in
  a renamed file)
- `pnpm docs:api`: clean regen (no static-method refs in
  AccountManager.mdx)
- `pnpm build:docs`: clean; /docs/upgrading/0.3.0 route generated
- `pnpm test:example` (Tier 2): 9 passing in 1m
- Grep static-call sites in src/: 0

Release sequencing (post-2026-06-22)
------------------------------------

1. Push branch + open M9.4 PR → develop
2. Open develop → main batch PR
3. Open release/0.3.0 PR (version bump + CHANGELOG rename)
4. Create GitHub Release v0.3.0 → publish.yml fires → npm publish
   with SLSA v1 provenance via OIDC

Closes #213
Closes #270
