+++ CHANGELOG.md
@@ -7,6 +7,75 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+M9.4 — the BREAKING removal of the `AccountManager` static facade that
+was deprecated in 0.2.7. Closes the M9 milestone (#270) and the F8
+audit finding (#213). Ships in 0.3.0 once the ≥1 month deprecation
+window per MAINTENANCE.md §6 elapses (earliest 2026-06-22).
+
+### Breaking
+
+- `AccountManager` static methods (`AccountManager.createAccount`,
+  `AccountManager.getLabeledAccounts`, all 17 static methods) have
+  been removed. The deprecation warning shipped in 0.2.7 is now a
+  hard `TypeError: AccountManager.X is not a function`. Per
+  MAINTENANCE.md §6 the one-minor-release deprecation window
+  (≥1 month) elapsed before this release. Migration: replace
+  `AccountManager.X(...)` with `harness.runtime.accountManager.X(...)`
+  for advanced operations, or use the shortcut
+  `harness.accounts.<label>` for labeled-account access (the common
+  case — covers ~90% of test code). Full migration guide with
+  before/after snippets at
+  [/docs/upgrading/0.3.0](https://movehat.org/docs/upgrading/0.3.0).
+  Closes #213, #270.
+
+### Removed
+
+- `_resetDeprecationWarnings` test-only escape hatch from
+  `packages/movehat/src/core/AccountManager.ts` — no warnings remain
+  to dedup.
+- `AccountManager.deprecation.test.ts` — pinned the warning surface
+  that no longer exists (~100 LoC).
+- `AccountManager.global-state.test.ts` — the F8 pinning block pinned
+  the static-state contract that no longer exists. The per-instance
+  isolation tests survive in the new
+  `AccountManager.instance-isolation.test.ts` (~110 LoC removed, 4
+  isolation tests preserved).
+
+### Changed
+
+- `AccountManager.ts` class JSDoc rewritten — removed the "Two API
+  surfaces are exposed" structure and the "Static facade
+  (deprecated)" paragraph. The class now documents a single API
+  surface: the instance constructor + instance methods.
+- `AccountManagerOptions.poolPath` JSDoc trimmed — removed the
+  paragraph describing the static facade's eager-cwd capture which
+  no longer applies.
+- `runtime.test.ts` `beforeEach`/`afterEach` dropped the now-redundant
+  `AccountManager.clearPool()` calls — the runtime owns its own
+  AccountManager instance per invocation, so there's nothing to clear
+  between tests.
+- `AccountManager.test.ts` rewritten to use `new AccountManager()` per
+  describe block instead of the removed static API. Each describe
+  block instantiates a fresh manager in `beforeEach`; tests that
+  previously chained static save → static clear → static load now
+  use two instances (`mgr` and `restored`) to prove the round-trip
+  works across boundaries.
+
+### Docs
+
+- New page `/docs/upgrading/0.3.0` — migration guide covering the
+  common labeled-account case, advanced direct-manager calls,
+  standalone (non-Harness) usage, per-instance pool semantics, the
+  `process.chdir` behavior change, and a 4-question FAQ. Linked from
+  the breaking-change entry above.
+- `packages/docs/content/docs/meta.json` adds a new "Upgrading"
+  category between "API Reference" and "Community" so the guide is
+  reachable from the sidebar.
+- `ROADMAP.md` — backfilled the M9 milestone section between M8 and
+  M7 with the standard closeout table (M9.1 → M9.4 sub-PRs).
+
+## [0.2.7] - 2026-05-22
+
 Post-0.2.6 polish batch — 6 audit-pass / cleanup items closed without
 public API change. All additive or docs-only; shipped to `main` in
 PR #274 on 2026-05-21. Backfilled 2026-05-22 when the per-PR CHANGELOG
