+++ packages/docs/content/docs/upgrading/meta.json
@@ -1,0 +1,4 @@
+{
+  "title": "Upgrading",
+  "pages": ["0.3.0"]
+}
